Server is able to verify login and pull messages and contacts
The server is now able to verify the login details and pull the messages and the contacts. While pulling and logging in, you receive an object that consists of two fields, contacts and messages that are both arrays of objects.
contacts are sorted by the time of their last time of sending a message.