Contacts are now updating well and according to the last message 's time
In the previous versions, the contacts used to update based on the sender of the message only, meaning that if I had sent a message to some new, hew would not appear as a chat. Now, if I send a message to someone it is treated as the last message from that person and therefor, will be up top.