Chat is much better
The chat now automatically scrolls to the end when entering a chat.
The messages are now two different color for better UX.
Time is now displayed correctly and will stick to the top once going over it so you can know what was the date of the messages.
There are various optimizations that can be made in the application but I prefer to fix bugs and opt-in new necessary features, and only then I will make these optimizations.